Added city, sex and age generators
Added and fixed generators for city, sex and age, so the city is generated first, and then sex is generated according to each city's distribution, as well as the age is according to each city's distribution and sex. Pre-computed some values used for that, as their uses already take too long, and added their files.